pyproject.toml
[project]
name = "symphony-cc"
version = "0.1.0"
description = "Claude Code 특화 자율 오케스트레이션 서비스"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
symphonyctl = "lib.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["lib*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
